Absorb uptime code into shared golang structure, remove uptime CI build
