+++ .circleci/src/commands/@docker-commands.yml
@@ -53,7 +53,6 @@ docker-tag-images:
             trending-challenge-rewards
             crm
             trpc
-            uptime
             backfill-audio-analyses
             verified-notifications
           )
@@ -89,7 +88,8 @@ docker-tag-images:
               )
               ;;
             *)
-              false
+              echo "Unhandled service, updating no images."
+              images=()
               ;;
           esac
