+++ .circleci/src/workflows/creator.yml
@@ -15,13 +15,6 @@ jobs:
       name: test-mediorum-unittests
       context: Vercel
       service: mediorum-unittests
-  - push-docker-image:
-      name: push-uptime
-      context: [Vercel, dockerhub]
-      service: uptime
-      filters:
-        branches:
-          only: main
 
   # Deploy audius-protocol `main` branch (stage)
   - deploy-stage-nodes:
@@ -30,7 +23,6 @@ jobs:
         - test-mediorum
         - test-mediorum-unittests
         - push-mediorum
-        - push-uptime
       filters:
         branches:
           only: main
