+++ .circleci/src/workflows/discovery.yml
@@ -98,13 +98,6 @@ jobs:
       filters:
         branches:
           only: main
-  - push-docker-image:
-      name: push-uptime
-      context: [Vercel, dockerhub]
-      service: uptime
-      filters:
-        branches:
-          only: main
 
   - lint-discovery-provider:
       name: lint-discovery-provider
@@ -141,7 +134,6 @@ jobs:
         - push-pedalboard-mri
         - push-comms
         - push-trpc
-        - push-uptime
 
       filters:
         branches:
