+++ pkg/registrar/registrar.go
